requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
astrapy==0.6.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.7.0
cassandra-driver==3.29.0
cassio==0.1.5
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.38.0
dataclasses-json==0.6.4
datasets==2.14.7
deprecation==2.1.0
dill==0.3.7
distro==1.9.0
dnspython==2.4.2
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-RESTful==0.3.10
flask-restplus==0.13.0
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
frozenlist==1.4.1
fsspec==2023.10.0
geomet==0.2.1.post1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.25.2
huggingface-hub==0.20.3
hyperframe==6.0.1
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langchain-openai==0.0.7
langsmith==0.1.8
MarkupSafe==2.1.3
marshmallow==3.20.2
MouseInfo==0.1.3
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.12.0
openpyxl==3.1.2
orjson==3.9.15
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pinecone-client==3.1.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
PyAutoGUI==0.9.54
pycryptodome==3.20.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyGetWindow==0.0.9
PyJWT==2.8.0
pymongo==4.6.2
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pypdf==4.0.2
PyPDF2==3.0.1
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.30.2
regex==2023.12.25
reportlab==4.0.8
requests==2.31.0
rpds-py==0.12.0
rubicon-objc==0.4.7
sendgrid==6.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starkbank-ecdsa==2.2.0
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.2.1
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
